fixed #12 change quality type

+++ CoreModule/STRUCTURES/include/Element.h
@@ -49,7 +49,7 @@ class Element
         string type;
         Label label;
         bool to_draw = true;
-        int quality = 0;
+        double quality = 0.0;
 
     public:
         Element(vector<Point3D> * points, string type, Label label)
@@ -71,8 +71,8 @@ class Element
         Label get_label(){ return label; }
         bool is_drawable(){ return to_draw; }
         void set_draw_flag(bool to_draw) { this -> to_draw = to_draw; }
-        void set_quality(int quality){ this -> quality = quality; }
-        int get_quality(){ return this -> quality; }
+        void set_quality(double quality){ this -> quality = quality; }
+        double get_quality(){ return this -> quality; }
 };
 
 #endif // ELEMENT_H
